Code improvements (#28)
* Reformatted project
This excludes MycroftMouth to prevent merge conflicts with the current refactor of MycroftMouth

* Changed constructors to initializer lists
The pinMode line was removed since the same thing is always done in the ClickEncoder constructor anyways

* Changed traditional include guards to #pragma once
This is because it is supported by arduino and nearly all compilers and is much less messy

* Changed angle bracket to quotes
After researching I've realised that anything in a local library is still considered local and thus should use quotes. Since  is a system library it should use angle brackets

* Removed unused includes in headers and changed to forward declarations

* Changed pointers to references

* Decreased scope of variable and removed auto to be more explicit

* Replaced TimerOne with CC licensed alternative

* Added license and readme to ClickEncoder
The readme is there to help give info to anyone using the library

* Added HT1632 license and readme

* Removed local pgmspace.h in preference to system one
There's not much reason to keep a local version

* Added include guard to MouthImages.h

* Set clicksPerNotch to 2 to fix input bug
Previously a single turn of the knob would register two turns in the ClickEncoder. Passing 2 as the last parameter eliminates this problem by processing the second one out.

* Added extra dependency to TimerOne library

* Removed unnecessary import