Created MycroftComponent base class (#46)
Although simple, this is helpful to provide anonymous access to components for initialization and updating. Ultimately I would consider the following all valid components:

    MycroftArduino
    MycroftEyes
    MycroftEncoder
    MycroftMatrix (a container for MycroftWeather and MycroftMouth)

The update function should ideally be called every millisecond although at the moment due to the delay in MycroftEyes and MycroftMouth, it does not get called every millisecond. Making MycroftEncoder implement MycroftComponent and use the update function for the isr would eliminate the need for TimerOne library with the CPU interrupt.