This is related to issue #70
The upload.sh script has been changed in several ways:
- When run by default, it will now stop and restart the mycroft-enclosure-client service (use -s or -skipService to skip this behavior)
- When run by default, the upload will not occur if not needed (see below for what that means).  Use -f or -force to force an upload.
- After a successful update, the Arduino is queried over the serial line for its version number, and that self-report version is stored in the file /opt/enclosure/build/version.txt

The upload.sh script calls verifyArduino.py to check version info.  If there is a version.txt file, the value is compared from the self-reported version the Arduino responds with across the serial line.  However, if the .hex is newer (file mode date) than the stored version.txt it is considered invalid.

TODO: This code is based on the system.ping responding with a version number.  It needs to be updated to operate against the newly implemented system.version command of #69.