Implemented auto-detect of platform in verifyArduino.py
This uses the standard (using JSON XPath notation) of:
   /enclosure/platform = "mycroft_mark_1"
For the first reference Mycroft device.  When run on any other hardware
(e.g. a generic Pi3) it will result in:
   /enclosure/platform = "unknown"
This value is stored in the mycroft.configuration.SYSTEM_CONFIG

This also sets the default port, rate and timeout for the Mark 1 device.